requirements.txt
django==4
djangorestframework
drf_yasg
setuptools
psycopg2-binary
python-dotenv
django-cors-headers
requests
django-filter